Replace indexed based lookup with dictionary lookup
This is to address a performance issue that was seen when uploading
large metagenomes.  The FastaGFF uploader using an index call to
get the contig_length.  This winds up being slow when there are
10M plus contigs.  Fortunately there was already a dict of the
same information.  I'm not sure why that wasn't used in the first
place.  Testing shows this gives linear performance even for
very large metagenomes.